# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SARTU API",
        "description": "Documentación OpenAPI mínima",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/usuarios": {
            "get": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Listado de usuarios (vista)",
                "operationId": "856e795f5d5450966d6b2da23cad7554",
                "responses": {
                    "200": {
                        "description": "HTML view"
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Crear usuario (form submission)",
                "operationId": "649b8ee4cb315de3c09d4f93fcc64588",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "rol_id": {
                                        "type": "integer"
                                    },
                                    "empresa_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirect (HTML)"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
        "/usuarios/{usuario}/edit": {
            "get": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Editar usuario (vista)",
                "operationId": "beb714d603a9c52f617e336352557ba4",
                "parameters": [
                    {
                        "name": "usuario",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HTML view"
                    }
                }
            }
        },
        "/usuarios/{usuario}": {
            "put": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Actualizar usuario (form submission)",
                "operationId": "c6692ad8a9294473c96eb8ad5f76cc15",
                "parameters": [
                    {
                        "name": "usuario",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "rol_id": {
                                        "type": "integer"
                                    },
                                    "empresa_id": {
                                        "type": "integer"
                                    },
                                    "activo": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirect (HTML)"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Eliminar usuario",
                "operationId": "5c70f37950d3332311fc71d1fca928ac",
                "parameters": [
                    {
                        "name": "usuario",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect (HTML)"
                    }
                }
            }
        }
    },
...
}